requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiomysql==0.2.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
better-profanity==0.7.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
databases==0.9.0

fastapi==0.115.8

fonttools==4.55.3
frozenlist==1.5.0
git-filter-repo==2.47.0
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
multidict==6.1.0
mysql-connector-python==9.1.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.2.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PyMySQL==1.1.1
pyngrok==7.2.3
pyparsing==3.2.1
python-dateutil==2.9.0.post0

pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlparse==0.5.3
starlette==0.45.3
tqdm==4.67.1

typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0
Werkzeug==3.1.3
yarl==1.18.3
